+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 # 0.1.4
 * Changed Error Board hints to be more informative.
 * Changed key controls for Error Board.
-* Changed displaying error of incorrect command option to be more strict.
+* Changed error message format, when incorrect command option is put, to be more strict.
 # 0.1.3 (2024-01-31)
 * Added logging monitor resolution to `runtime.log`.
 * Added highlighting of in-game exp tooltip region for debug mode.