+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 * Changed Error Board hints to be more informative.
 * Changed key controls for Error Board.
 * Changed error message format, when incorrect command option is put, to be more strict.
+* Fixed temporal displacement in debug mode.
 # 0.1.3 (2024-01-31)
 * Added logging monitor resolution to `runtime.log`.
 * Added highlighting of in-game exp tooltip region for debug mode.